refactor: changed dataType from int to double

+++ common/src/main/java/et/edu/astu/common/dto/DepositRequest.java
@@ -1,3 +1,3 @@
 package et.edu.astu.common.dto;
 
-public record DepositRequest(long accountNumber, int amount) { }
+public record DepositRequest(long accountNumber, double amount) { }
